Moved non fips tests to rest-utils/core and kep FIPS in fips-tests.
